edited primitives as per requested

+++ src/main/java/Primitives.java
@@ -1,6 +1,6 @@
 public class Primitives {
-    int b = 65;
-    int s = 65;
+    byte b = 65;
+    short s = 65;
     int i = 65;
     long l = 65L;
     float f = 65f;
